requirements.txt
matplotlib==3.9.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
streamlit==1.38.0
tensorflow==2.17.0
urllib3==2.2.3

